pyproject.toml
[project]
name = "oasis-llm"
version = "0.1.0"
description = "OASIS image rating with LLM agents"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.52",
    "duckdb>=1.1",
    "numpy>=2.0",
    "pydantic>=2.9",
    "typer>=0.12",
    "rich>=13.9",
    "python-dotenv>=1.0",
    "tenacity>=9.0",
    "pyyaml>=6.0",
    "streamlit>=1.39",
    "pandas>=2.2",
    "matplotlib>=3.9",
    "pillow>=11.0",
    "httpx>=0.27",
    "ollama>=0.4",
    "langfuse<3",
    "streamlit-autorefresh>=1.0.1",
]

[project.scripts]
oasis-llm = "oasis_llm.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/oasis_llm"]
